Refresh the sensor cache when the cache vessel is changed
This only happens initially when the sensor variables (s.*) are accessed - and the sensors are not
updated until the next frames update, so the first time querying the sensors will return an empty response.
This patch just triggers an internal event when the Vessel is changed on the ModuleCache, ensuring that when
the sensor data is called for it has already been scanned.