Implement a Plugin system
This lets third-party mods supply their own API data without a dependency on Telemachus.
To do so, a class that matches the form of IMinimalTelemachusPlugin (but is not required
to be explicitly inherited from it or implement it) should be passed to
Telemachus.PluginRegistration.Register, easily found and invoked via reflection.

The plugin classes can specify exact or wildcard API strings, and is lazily evaluated for
delegate functions for each of the API strings that it handles.