Scan for MonoBehaviours with a Telemachus registration action
If any of the MonoBehaviours have a "public static Action<object> TelemachusPluginRegister" property, that
property will be set to the PluginRegistration::Register method, to allow painless registration for compatible
mods.

This avoids all the messy business with reflection on the mod end of thing. It also e.g. explicitly
trying to instantiating anything user-side, or making assumptions about what they want to register.